fix(component) ensure usage of default strategy name constant, move duration selector.
